requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
absl-py=0.11.0=pyhd3eb1b0_1
astor=0.8.1=py36haa95532_0
blas=1.0=mkl
brotlipy=0.7.0=py36h2bbff1b_1003
ca-certificates=2021.1.19=haa95532_0
certifi=2020.12.5=py36haa95532_0
cffi=1.14.5=py36hcd4344a_0
chardet=4.0.0=py36haa95532_1003
click=7.1.2=pyhd3eb1b0_0
console_shortcut=0.1.1=4
coverage=5.4=py36h2bbff1b_2
cryptography=3.3.1=py36hcd4344a_1
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cython=0.29.21=py36hd77b12b_0
flask=1.1.2=pyhd3eb1b0_0
gast=0.2.2=py36_0
google-pasta=0.2.0=py_0
google-trans-new=1.1.9=pypi_0
grpcio=1.35.0=py36hc60d5dd_1
h5py=2.10.0=py36h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
idna=2.10=pyhd3eb1b0_0
importlib-metadata=2.0.0=py_1
intel-openmp=2020.2=254
itsdangerous=1.1.0=py36_0
jinja2=2.11.3=pyhd3eb1b0_0
joblib=1.0.1=pyhd3eb1b0_0
keras-applications=1.0.8=py_1
keras-base=2.3.1=py36_0
keras-gpu=2.3.1=0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
libprotobuf=3.14.0=h23ce68f_0
markdown=3.3.3=py36haa95532_0
markupsafe=1.1.1=py36he774522_0
mkl=2020.2=256
mkl-service=2.3.0=py36h196d8e1_0
mkl_fft=1.2.1=py36h46781fe_0
mkl_random=1.1.1=py36h47e9c7a_0
nltk=3.5=py_0
numpy=1.19.2=py36hadc3359_0
numpy-base=1.19.2=py36ha3acd2a_0
openssl=1.1.1j=h2bbff1b_0
opt_einsum=3.1.0=py_0
pandas=1.1.3=py36ha925a31_0
pip=21.0.1=py36haa95532_0
protobuf=3.14.0=py36hd77b12b_1
pycparser=2.20=py_2
pyopenssl=20.0.1=pyhd3eb1b0_1
pyreadline=2.1=py36_1
pysocks=1.7.1=py36haa95532_0
python=3.6.12=h5500b2f_2
python-dateutil=2.8.1=pyhd3eb1b0_0
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.4.1=py36h2bbff1b_1
regex=2020.11.13=py36h2bbff1b_0
requests=2.25.1=pyhd3eb1b0_0
scikit-learn=0.23.2=py36h47e9c7a_0
scipy=1.5.2=py36h9439919_0
setuptools=52.0.0=py36haa95532_0
six=1.15.0=py36haa95532_0
sqlite=3.33.0=h2a8f88b_0
tensorboard=2.0.0=pyhb38c66f_1
tensorflow=2.0.0=gpu_py36hfdd5754_0
tensorflow-base=2.0.0=gpu_py36h390e234_0
tensorflow-estimator=2.0.0=pyh2649769_0
tensorflow-gpu=2.0.0=h0d30ee6_0
termcolor=1.1.0=py36haa95532_1
threadpoolctl=2.1.0=pyh5ca1d4c_0
tqdm=4.56.0=pyhd3eb1b0_0
urllib3=1.26.3=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=0.16.1=py_0
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py36haa95532_0
wincertstore=0.2=py36h7fe50ca_0
wrapt=1.12.1=py36he774522_1
yaml=0.2.5=he774522_0
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
